Fixes Old OS Crash
Changed a lot of the code.

Map is now a fragment of a blank activity. This gives us a bar on top
with the app name. I also implemented an overflow menu so the user can
select Satellite/Normal/etc. I just did that cause I could, but we could
put settings in there instead.  Also made the marker purple for current
location.

App runs on my Nexus 4 so could run on older OSes.